requirements.txt
streamlit>=1.28.0
pycryptodome>=3.19.0

